backend/requirements/02-core.txt
#
# Please keep this file alphabetical
#
analytics-python==1.4.0       # segment analytics wrapper for Python
avalara==21.12.0              # Avalara (Tax Nexus) SDK for Python
boto3==1.19.5                 # The AWS SDK for Python
celery==5.1.2                 # Distributed Task Queue
celery_once==3.0.1            # Allows you to prevent multiple execution and queuing of celery tasks
dataclasses-json==0.5.6       # Easily convert dataclasses to JSON and dictionary types
Django==3.2                   # Django web framework
django-anymail==8.4           # Email backend for django and ESPs
django-celery-beat==2.2.1     # Database-backed Periodic Tasks
django-celery-results==2.2.0  # Celery result backends for Django
django-chunkator==2.0.0       # Chunk Django queries
django-dirtyfields==1.7.0     # Detect object-level changes
django-environ-2==2.3.0       # Environment variables for Django
django-extensions==3.1.3      # Convenient extensions for django
django-fsm==2.7.1             # Django friendly finite state machine support
django-fsm-admin==1.2.4       # Integrate django-fsm state transitions into the django admin
django-grappelli==2.15.3      # Drop-in theme for django admin
django-health-check==3.4.1    # Monitoring for resources
django-hijack==3.1.4          # django-hijack allows superusers to hijack (=login as) and work on behalf of another user.
django-hijack-admin==2.1.10
django-json-widget==1.1.1     # Django json widget is an alternative widget that makes it easy to edit the jsonfield field of django
django-localflavor==3.1       # Country-specific Django helpers
django-model-utils==4.1.1     # Model utilities for Django
django-redis==4.12.1          # Full featured redis cache backend for Django.
django-robots>=4.0.0          # Robots exclusion application for Django
django-simple-history==3.0.0  # Object historical tracking for django models
django-soft-delete==0.9.18    # Soft Delete model instances
django-storages==1.12.2       # Django Amazon S3 file storage
djangorestframework==3.12.4   # RESTful API framework for Django
djangorestframework-camel-case==1.2.0
flower==1.0.0                 # Web based tool for monitoring and administrating Celery clusters
furl==2.1.3                   # parsing and manipulating URLs easy
gunicorn==20.0.4 # Server to run app in production
newrelic==7.6.0.173           # New Relic Python Agent
python-dateutil==2.8.2        # Convenience utilities for dates with Python
ShopifyAPI==9.0.0             # Shopify API for Python
stripe==2.61.0                # Python bindings for the Stripe API
whitenoise==5.3.0             # Radically simplified static file serving for WSGI applications
sentry-sdk==1.4.3             # Sentry error monitoring
supervisor==4.2.2             # Manager for gunicorn and other processes
stripe==2.61                  # Library for interfacing with stripe api
requests-mock==1.9.3        # HTTP request mock
